backend/requirements.txt
# ==================== Core Framework ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ==================== Authentication & Security ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# ==================== Database ====================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# ==================== Configuration & Validation ====================
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# ==================== AI Providers (Cascade Fallback System) ====================
# OpenAI (GPT-4, GPT-3.5)
openai>=1.6.1

# Anthropic (Claude 3.5 Sonnet)
anthropic>=0.7.0

# Vector Database
chromadb==0.5.23

# LangChain (optional enhanced features)
langchain==0.0.340
langchain-community==0.0.6
langchain-openai==0.0.2

# ==================== Document Processing ====================
pypdf==3.17.1
python-docx==1.1.0
aiofiles==23.2.1

# ==================== Data Processing ====================
pandas==2.1.4
numpy>=1.26.0

# ==================== Real-Time Communication ====================
python-socketio==5.10.0

# ==================== Caching & Background Tasks ====================
redis==5.0.1
celery==5.3.4

# ==================== Utilities ====================
requests==2.31.0